--- delivery/src/main/java/com/sparta/msa/delivery/application/dto/DeliveryDTO.java
@@ -1,4 +1,0 @@
-package com.sparta.msa.delivery.application.dto;
-
-public class DeliveryDTO {
-}
