+++ delivery/src/main/java/com/sparta/msa/delivery/presentation/request/UpdateDeliveryRequest.java
@@ -1,6 +1,6 @@
 package com.sparta.msa.delivery.presentation.request;
 
-import com.sparta.msa.delivery.application.dto.UpdateDeliveryDto;
+import com.sparta.msa.delivery.application.dto.delivery.UpdateDeliveryDto;
 import com.sparta.msa.delivery.domain.model.DeliveryStatus;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Size;
